simplify useApiMutation logic and extract types
